Add data model to track API requests
This commit introduces a data model to store and persist API requests. This decision was made with simplicity in mind, so that other developers can start this project up on their local machine with the simplest possible instructions:

1. `pip install -r requirements.txt`
2. `./manage.py runserver`

One could make a case that using the database (even an in-memory one) to store things like the count of requests and the most recent request time is unnecessary overhead. I would tend to agree, and as we look to scale this project up, it may make sense to introduce redis as a cache backend so we can avoid hitting the db on every request.